massive misc features update, see desc for details
- you can now clean yourself with space cleaner and water
- adds an optimization to some typechecks
- removes CRASH_SAFE an replaces it with either CRASH or log_error
- fixes some dialogue errors
- fixes bloody footstep logic
- fixes blood overlays applying to the whole body
- tweaks projectile collision logic and priority

+++ code/_core/datum/reagent/reagent_core.dm
@@ -270,14 +270,19 @@
 
 	. = ..()
 
-	if(. && isturf(target))
-		var/turf/T = target
-		var/cleaning_power = volume_to_splash*10*strength_mod
-		for(var/obj/effect/cleanable/C in T.contents)
-			if(cleaning_power <= 0)
-				break
-			qdel(C)
-			cleaning_power -= 10
+	if(.)
+		if(isturf(target))
+			var/turf/T = target
+			var/cleaning_power = volume_to_splash*10*strength_mod
+			for(var/obj/effect/cleanable/C in T.contents)
+				if(cleaning_power <= 0)
+					break
+				qdel(C)
+				cleaning_power -= 10
+		else if(is_advanced(target))
+			var/cleaning_power = volume_to_splash*strength_mod
+			var/mob/living/advanced/A = target
+			A.make_clean(clean_strength=cleaning_power)
 
 /reagent/phenol
 	name = "Phenol"
